Adjust inline comment syntax to conform to JSON
JSON does not support `//`-style comments.  For the sake of string
search, the TODOs and NOTEs were put into namespaces that can be cut
from the examples in the future when issues are resolved or otherwise
annotated sufficiently.

With these changes applied, the Bash one-liner from `d2d218e` runs
successfully.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>